app-code/bkd/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
mysql-connector-python==8.0.15
python-dotenv==1.0.1